# src/main/resources/swagger/swagger-api.json
{
    "swagger": "2.0",
    "info": {
      "title": "Trouble Ticket",
      "description": "This is Swagger UI environment generated for the TMF Trouble Ticket specification",
      "version": "2.0"
    },
    "host": "serverRoot",
    "basePath": "/tmf-api/troubleTicket/v2/",
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/hub": {
        "post": {
          "operationId": "registerListener",
          "summary": "Register a listener",
          "description": "Sets the communication endpoint address the service instance must use to deliver information about its health state, execution state, failures and metrics.",
          "tags": [
            "events subscription"
          ],
          "parameters": [
            {
              "name": "data",
              "schema": {
                "$ref": "#/definitions/EventSubscriptionInput"
              },
              "required": true,
              "in": "body",
              "description": "Data containing the callback endpoint to deliver the information"
            }
          ],
          "responses": {
            "201": {
              "description": "Subscribed",
              "schema": {
                "$ref": "#/definitions/EventSubscription"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/hub/{id}": {
        "delete": {
          "operationId": "unregisterListener",
          "summary": "Unregister a listener",
          "description": "Resets the communication endpoint address the service instance must use to deliver information about its health state, execution state, failures and metrics.",
          "tags": [
            "events subscription"
          ],
          "parameters": [
            {
              "name": "id",
              "type": "string",
              "required": true,
              "in": "path",
              "description": "The id of the registered listener"
            }
          ],
          "responses": {
            "204": {
              "description": "Deleted"
            },
            "405": {
              "description": "Method not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "400": {
              "description": "Bad request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      }
    },
    "definitions": {
...
      "EventSubscription": {
        "required": [
          "id",
          "callback"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the listener"
          },
          "callback": {
            "type": "string",
            "description": "The callback being registered."
          },
          "query": {
            "type": "string",
            "description": "additional data to be passed"
          }
        }
      },
      "EventSubscriptionInput": {
        "required": [
          "callback"
        ],
        "properties": {
          "callback": {
            "type": "string",
            "description": "The callback being registered."
          },
          "query": {
            "type": "string",
            "description": "additional data to be passed"
          }
        }
      },
      "Error": {
        "required": [
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "description": "Application related code."
          },
          "reason": {
            "type": "integer",
            "description": "Text that explains the reason for error."
          },
          "message": {
            "type": "string",
            "description": "(optional) Text that provide more details and corrective actions related to the error."
          },
          "status": {
            "type": "integer",
            "description": "(optional) http error code extension like 400-2"
          },
          "referenceError": {
            "type": "string",
            "description": "(optional) A URL to online documentation that provides more information about the error."
          },
          "@type": {
            "type": "string",
            "description": "(optional) The class type of a REST resource."
          },
          "@schemaLocation": {
            "type": "string",
            "description": "(optional) A link to the schema describing a REST resource."
          }
        }
      }
    }
  }